requirements.txt
botbuilder-integration-aiohttp>=4.14.0
azure-cosmos==4.2.0
marshmallow==3.15.0
aiohttp~=3.8.1
simplejson==3.17.6
marshmallow-dataclass==8.5.8
stringcase==1.2.0
